+++ src/KomTracker/KomTracker.Infrastructure/Mail/Brevo/Models/ChangeEmailConfirmationTemplateParamsModel.cs
@@ -1,0 +1,6 @@
+﻿namespace KomTracker.Infrastructure.Mail.Brevo.Models;
+internal class ChangeEmailConfirmationTemplateParamsModel
+{
+    public string firstName { get; set; }
+    public string url { get; set; }
+}